Volumes: cycles rendering support of volume object as dense grids
This does not yet use OpenVDB grid transforms properly, in particular
it is assumed all grids have the same transform and do not contain
rotation or shear.

The conversion is also not optimized. Grid bounds are being computed
multiple times, and for the bounding mesh we should be using the VDB
tree directly.