Objects: modifier stack integration for hair, point clouds and volumes
This is more a prototype for testing, it's not clear yet if we want to use
the existing modifier stack or immediately go to nodes.

Currently working are a few deform modifiers for hair and points clouds.
The infrastructure for constructive modifiers is there as well, just none
are implemented.