Cleanup: Animation, simplified fcurve_eval_between_keyframes() code
The function now immediately returns the evaluated value, rather than
setting a variable's value and returning that at the end of the
function. As a result, `else`-clauses and `break` statements could be
removed, simplifying and unindenting the code.

Declarations of some variables that are only used in a specific part of
the function have been moved there to clarify and limit their scope.

No functional changes.