Initial demo step for IDType refactor 'cleanup' project.
This is mostly a first version, consider it a base of discussion for how we want to do it.

Once this is agreed on, I'll add all missing callbacks, then on cleanup day I'll add usage of those callbacks in `lib_id` idmanagement code, and me and others can go monkey-editing all our ID types BKE files accordingly.

Regarding `BKE_idcode.h`/`idcode.c`, I think these should then be merged in `BKE_idtype`, then we can see which of idcode functions we keep as-is (as wrappers around new IDTypeInfo), and which we can simply remove in favour of direct usage of IDTypeInfo data by calling code?

This is part of T73719.

Differential Revision: https://developer.blender.org/D6966