Allows to preserve UUIDs on copy, outside of copy-on-write
Could be useful for clipboard implementation where, for example, strips
live outside of any reachable database. The UUID will be generated on
paste, so there is no need to generate new UUID on copy.

While generating UUID is almost harmless, incrementing global counter
increases chance of overflowing the integer sooner.

This changes makes it possible to solve issue when user opens .blend
file with strips, which will generated UUIDs in the lower part of the
integer, and then holds Ctrl-C for a very long time.