Materials: add custom object properties as uniform attributes.
This patch allows the user to type a property name into the
Attribute node, which will then output the value of the property
for each individual object, allowing to e.g. customize shaders
by object without duplicating the shader.

In order to make supporting this easier for Eevee, it is necessary
to explicitly choose whether the attribute is varying or uniform
via a dropdown option of the Attribute node. The Cycles design
seems to intend to treat all attributes equally, so the Blender
interface uses a name prefix that can't be entered with keyboard.

Differential Revision: https://developer.blender.org/D2057