Materials: support the custom uniform attributes in Eevee.
The attributes are provided to the shader via a UBO indexed with
resource_id, similar to the existing Object Info data. Unlike that,
however, it is necessary to maintain a separate buffer for every
requested combination of attributes.

This is done using a hash table with the attribute set as the key,
as it is not inconceivable that technically different materials may
use the same set of attributes. In addition, in order to minimize
wasted memory, a sparse UBO pool is implemented, so that chunks that
don't require that set of data don't have to allocate any memory.

Differential Revision: https://developer.blender.org/D2057