USD: remove library initialisation hack
Remove the hack for library initialisation; this is no longer necessary
as the required information can be passed to the USD library after its
static initialisers have run.

This means that USD's static initialisers will not be able to find the
USD plugin files. This is fine, as Blender now provides it with the
correct path before the first use of the library.

The removed hack was two-fold. It consisted of patching the USD library
to add a function, and calling that function from Blender. Now with that
function call removed from Blender's code, everything will keep working
regardless of whether the USD library is still patched or not. This'll
make bisecting things in the future a bit easier.

Manifest Task: https://developer.blender.org/T80320