Animation: move group colors switch to user preferences
Move the "Show Group Colors" toggle from a per-editor option to a single
user preference in the Animation preferences.

The old "Show Group Colors" toggle had to be set per editor, and was on
by default. This meant that disabling group colors would require an action
for every file, for every editor.

Differential Revision: https://developer.blender.org/D9391