Transform: New feature to set a custom 'Snap With'
Ref T66484

Basically the idea is to use a modal keymap to change the current Base Point
defined by `Snap With` by one that can be chosen through snapping.

{F9170047}

I prefer to avoid using any of the modifier buttons (Shift, Alt, Ctrl) as
they can be used in the future to allow navigation during transform operations.

So the shortcut chosen in this patch is the {key B}.

Note:
This feature is not enabled if the scene snap option is Incremental or grid.

Maniphest Tasks: T66484

Differential Revision: https://developer.blender.org/D9415