Geometry Nodes: Support the "normal" attribute
The `normal` attribute is a bit special for a few reasons:
 - It has the special CD_NORMAL custom data type even though
   it is just a float3.
 - It is sometimes stored in the `MVert` array as a vector of
   shorts instead of in a separate custom data layer.
 - The custom data layer does not have a default name, so we
   can't count on addressing it by name.
 - It can exist on multiple domains, meaning that polys, vertices,
   and corners can all have the "normal" attribute.

All of these combined raise a lot of questions about how to properly
support reading and writing from this attribute. The whole idea that
it should be valid to address an attribute with the same name on
multiple domains is challenging for the attribute code.

This branch currently supports reading normals stored in the `MVert`
layer and in `CD_NORMAL` layers. I'm not at all sure that this is the
ideal way of dealing with this situation, but it's a start.