Start implementing attribute domain conversions
This doesn't compile, and it isn't even close to finished, but it's a
start at basic domain conversion, which are generally simple operations
that just depend on the topology of the mesh.

Next steps are creating a new attribute with the calculated
values, and more importantly templating this to work with
different attribute types.