Attribute Search: Always add current value to the search list
This way the button isn't red if the socket value is any of
"HELLO!", "DOES", "THIS", "WORK???".

And the answer is no, it doesn't work, because all buttons are still
not interactive. I haven't been able to figure out why, but the UI code
doesn't see them as active (i.e. they are deactivated before they can
open the menu).