Attribute Search: Fix interaction with buttons not working
When an "exec" function was set, the data argument was provided to
the button. This meant that two of these buttons won't be considered
"equal" over successive redraws, since the data is reallocated every time.
To maintain "active" status over redraws, that new-button to old-button
map needs to be valid,, and it wasn't so the buttons weren't active.

The take-away: search button arguments can not be reallocated for every
redraw if an exec function is provided.

Since I have since connected the search button to the "default_value"
RNA property, we don't actually need the manual "exec" function anyway,
so this commit just removes it.