Update addons_contrib in make.bat.
Invoking new update_nv_addons.cmd script in make.bat
to checkout and pull NVIDIA's branch of addons_contrib.
This is currently necessary to always include the latest
UMM addon in the build.  This is not an ideal solution
because we pull the latest commits from this branch
on every build, so we need to consider alternatives
to this approach.