USD Export: spurious error message.
Calling BLI_dir_create_recursive() with a path that has a
trailing slash generates an error message that the directory
couldn't be created, even though the creation is successful.
The error message appears to be spurious, but I changed the
code to use a path with no trailing slash, to avoid this.