USD export: consistent camera aperature export.
The current camera aperature export code computes
the aperature setting based on the camera sensor
and the scene's render resolution. This is not
consistent with how the camera sensor is set on
import and causes unintuitive behavior where the
camera export gives different results base on
the a given scene's render settings.  This needs
to be revisited, but, for now, setting the aperature
alues directly from the camera sensor components.