Zoom influence factor for edge scroll speed.
This allows modifying the scroll speed based on the View2D zoom factor.
With influence 0.0 (default behavior) the speed will be constant in
UI space as before.
At maximum influence 1.0 the speed will instead max out in the view
space, i.e. scrolls slower in UI space when zooming out.
At intermediate values the zoom influence is proportionally reduced,
for example at influence 0.5 the view will scroll as if zoom factor is
half of the actual zoom.

This feature is used in the node editor to slow down scrolling a bit
when zoomed out.