Attempt at fixing T88281: Pose Library 'flip pose' sometimes flips wrong
When flipping a pose in the pose library, sometimes the resulting pose
can be wrong. This seems to happen when the rest orientation of the bone
has its X-axis straight up/down. In these cases, an extra -1 scale to
the X and Z axes of the resulting matrix seems to fix things.

This is not a final fix, just a commit to a branch so that animators can
get a test build.