[WIP] VSE: Transform overwrite mode
WIP status: logic is implemented, and overwrite mode is currently forced.

Add mode to overwrite strips on overlap instead of resolving overlap.
When overlap is created, 3 things can happen:
 - On partial overlap, handles of overlapped strip are moved
 - On complete overlap with smaller strip, overlapped strip is split
 - On complete overlap with larger strip, overlapped strip is removed

There are some edge cases where this mode may not work very well, this
is mainly case for partial overlap with effects. This situation can be
resolved in more than one way and I don't think it can be predicted.

Differential Revision: https://developer.blender.org/D11805