USD Import: add cache file only if needed
Fixed bug where a cache file was always created
for the imported USD, even when the USD has no
animation. Modified the logic to add a cache file
only if a reader requires it. This is not only
more efficient, but also avoids the problem
where we can't overwrite the USD the cachefile is
referencing because it has a pointer to the
open stage for the lifetime of the scene.