Asset Catalogs: defer writing of catalog definitions
Instead of writing the catalog definition file (CDF) immediately after
creating a new catalog, wait until the `write_to_disk(...)` function is
called.

As a result, it's now also possible to create an `AssetCatalogService`
without passing it an asset library root path (for example when creating
catalogs in a not-yet-saved blend file and without having an asset
library opened/configured/selected).