UUIDs: rename UUID to bUUID
Rename the `UUID` struct & C++ class to `bUUID`. This avoids a symbol
clash on Windows, which also defines `UUID`. This only pops up when a
`UUID` field is used in the DNA code, which is why it wasn't a problem
before (it will be once D12589 lands).

No functional changes.