Asset Catalog Tests: better path separator handling
Fixes by @LazyDodo to make the tests run on Windows.

This requires another change to get working properly, namely updating
`BLI_dir_create_recursive()` on Windows to call `BLI_path_slash_native()`
on the path before trying to create it. This will be handled in a different
commit, as it's not catalog-related.