Geometry Nodes: Separate + Delete Geometry for fields
Delete Geometry:
This adds a copy of the old node in the legacy folder and updates the node to
work with fields. The invert option is removed, because it something that should
be very easy with fields, and to be consistent with other nodes which have a
selection. There is also a dropdown to select the domain, because the domain
can't be determined from the field input. When the domain does not belong on
any of the components an info message is displayed.

{F10416062}

Separate Geometry:
A more general version of the old Point Separate node. The "inverted" output is the
same as using the delete geometry node.

{F10518721}

Possible things to change:
- The name of the outputs of the Separate Geometry node
- Add option for "smooth" {D10979} . This is probably best for a dedicated node since it only applies to meshes.

Reviewed By: JacquesLucke

Differential Revision: https://developer.blender.org/D12574